Bug 25755: Add Koha::Item->request_transfer method
This patch adds the request_transfer method to the Koha::Item object.
The new method requires `to` and `reason` parameters and will throw an
exception if an active transfer request is found on the item.

Test plan
1/ Run the included updated tests in t/db_dependent/Koha/Item.t
2/ Verify the tests pass
3/ Signoff

TODO: Complete work of LIFO queues for get_transfer and add tests for
that