Refine GetTransfersFromTo to filter out unsent transfers
GetTransfersFromTo is used solely by the transferstorecieve page which
assumes items have already been sent. With the introduction of
daterequested we allow for a state where a transfer has been scheduled
but not yet sent (See stockrotation for such a case) and so we need to
filter these transfers out from the transferstorecieve report.