Bug 24631: Convert Koha::Plugins to a Koha::Objects class
This patch adds the plugins table to the database, updates the
Koha::Plugins object to be a Koha::Objects based class, adds the
corresponding Koha::Plugin class, including a new 'load_plugin' method.

It also moves the InstallPlugins method into it's own
Koha::Plugins::Installer class and renames it to 'refresh' to better
represent it's function.