Bug 36640: Remove fnAddFilters
Also centralize and clean the delay code to use DT's debounce.

I have commented the code from initComplete as I don't think it's the
right place. Maybe we need to add it for other table, but then it should
be added when we init DT.

Removed blacklist_keys to keep code clean, could be re-added later if we
really need it.